% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_proj.R
\name{as_proj}
\alias{as_proj}
\title{Create CRS definition in Proj system}
\usage{
as_proj(x)
}
\arguments{
\item{x}{A list of named Proj elements}
}
\value{
A length-1 character string with the Proj-style CRS definition
}
\description{
This function creates a new Proj-style CRS definition from list of named
elements.
}
\examples{
as_proj(
  list(
    proj = "omerc",
    lat_0 = 37.385,
    lonc = -108.51,
    alpha = 0,
    gamma = 0,
    k_0 = 1,
    x_0 = 500,
    y_0 = 500
  )
)
}
