% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_legacy_grid.R
\name{define_legacy_grid}
\alias{define_legacy_grid}
\title{Define a floating grid coordinate reference system for a legacy grid}
\usage{
define_legacy_grid(
  pt_0_geo,
  pt_0_grid = c(0, 0),
  pt_1_geo,
  pt_1_grid,
  name = NULL
)
}
\arguments{
\item{pt_0_geo}{A numeric vector of the geographic coordinates
(\code{c(longitude, latitude)}) of the primary datum,
in WGS84 decimal degrees}

\item{pt_0_grid}{A numeric vector of the geographic coordinates
(\code{c(easting, northing)}) of the primary datum,
in local coordinates (defaults to \code{c(0,0)})}

\item{pt_1_geo}{A numeric vector of the geographic coordinates
(\code{c(longitude, latitude)}) of the secondary point,
in WGS84 decimal degrees}

\item{pt_1_grid}{A numeric vector of the geographic coordinates
(\code{c(easting, northing)}) of the secondary point,
in local coordinates}

\item{name}{An optional name for the floating grid coordinate
reference system}
}
\value{
Object of class \code{crs}, which is a list with elements \code{input}
(length-1 character) and \code{wkt} (length-1 character).
}
\description{
This function defines a floating grid coordinate reference system in the
Oblique Mercator projection for a legacy (existing) floating grid
based on the coordinates for a primary datum and a secondary point.
}
\details{
This function requires the latitude and longitude coordinates
(in WGS84 decimal degrees) of the
primary datum and a secondary point.
}
\examples{
define_legacy_grid(
  pt_0_geo = c(-108.51, 37.385),
  pt_0_grid = c(500, 500),
  pt_1_geo = c(-108.511, 37.384),
  pt_1_grid = c(360, 380),
  name = "Screwy Legacy Grid"
)
}
