% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_new_grid.R
\name{define_new_grid}
\alias{define_new_grid}
\title{Define a floating grid coordinate reference system for a new grid}
\usage{
define_new_grid(pt_0_geo, pt_0_grid = c(0, 0), name = NULL)
}
\arguments{
\item{pt_0_geo}{A numeric vector of the geographic coordinates
(\code{c(longitude, latitude)}) of the primary datum,
in WGS84 decimal degrees}

\item{pt_0_grid}{A numeric vector of the geographic coordinates
(\code{c(easting, northing)}) of the primary datum,
in local coordinates (defaults to \code{c(0,0)})}

\item{name}{An optional name for the floating grid coordinate
reference system}
}
\value{
Object of class \code{crs}, which is a list with elements \code{input}
(length-1 character) and \code{wkt} (length-1 character).
}
\description{
#' This function defines a new floating grid coordinate reference system in the
Oblique Mercator projection for a new floating grid
based on the coordinates for a primary datum.
}
\details{
This function requires the latitude and longitude coordinates
(in WGS84 decimal degrees) of the
primary datum for a new floating grid.
}
\examples{
define_new_grid(
  pt_0_geo = c(-108.51, 37.385),
  pt_0_grid = c(500, 500),
  name = "Great New Grid"
)
}
